Don't invalidate everything for a search query
Searches use POST (rather than the more-common GET) request, and we
requery everything on mutations (by default) to ensure cache
consistency. However, this means that searching does a _ton_ of
revalidation.

Disable this behavior when searching .. and make some searchbox cleanups
along the way.
